tavern_server: add db_test macro
Tests using the database now only need to be decorated with the
`db_test` attribute:

```rust
\#[db_test]
async fn some_test_name() {
    // no need to set up the database
}
```
