+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "tavern_derive",
     "tavern_pathfinder",
     "tavern_server",
 ]
