+++ tavern_derive/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "tavern_derive"
+version = "0.1.0"
+authors = ["Michael Bryant <shadow53@shadow53.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+quote = "1.0"
+syn = "1.0"
+
+[lib]
+proc-macro = true
