+++ tavern_server/Cargo.toml
@@ -22,7 +22,7 @@ maintenance = { status = "actively-developed" }
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-test-db = [] # Used to conditionally enable db tests
+db-test = [] # Used to conditionally enable db tests
 
 [dependencies]
 base64 = "0.12"
@@ -38,6 +38,7 @@ serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sqlx = { version = "0.3.4", default-features = false, features = [ "chrono", "postgres", "uuid", "runtime-tokio" ] }
 structopt = "0.3"
+tavern_derive = { version = "0.1", path = "../tavern_derive" }
 tokio = { version = "0.2", features = ["full"] }
 uuid = { version = "0.8", features = ["v4", "serde"] }
 warp = "0.2"
